Relocate ontology(Actions|Queries|Objects) into their respective directories
