+++ packages/generator/src/v1.1/generateMetadataFile.test.ts
@@ -37,13 +37,13 @@ describe(generateMetadataFile, () => {
     ).toMatchInlineSnapshot(`
       "import type { OntologyDefinition } from '@osdk/api';
       import type { Ontology as ClientOntology } from '@osdk/legacy-client';
+      import type { Actions } from './ontology/actions/Actions';
       import { markTodoCompleted } from './ontology/actions/markTodoCompleted';
+      import type { Objects } from './ontology/objects/Objects';
       import { Person } from './ontology/objects/Person';
       import { Todo } from './ontology/objects/Todo';
+      import type { Queries } from './ontology/queries/Queries';
       import { getCount } from './ontology/queries/getCount';
-      import type { Actions } from './ontologyActions';
-      import type { Objects } from './ontologyObjects';
-      import type { Queries } from './ontologyQueries';
 
       export const Ontology: {
         metadata: {
